requirements.txt
langchain
langchain-community
faiss-cpu
scikit-learn
streamlit
python-dotenv
